+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/manipulators/PlaybackEventDataManipulator.kt
@@ -34,12 +34,7 @@ internal class PlaybackEventDataManipulator(
         }
 
         // isLive
-        data.isLive =
-            Util.getIsLiveFromConfigOrPlayer(
-                playbackInfoProvider.playerIsReady,
-                metadataProvider.getSourceMetadata()?.isLive,
-                exoPlayer.isCurrentMediaItemDynamic,
-            )
+        data.isLive = metadataProvider.getSourceMetadata()?.isLive ?: exoPlayer.isCurrentMediaItemDynamic
 
         // we report 0 videoDuration for live streams to be consistent with other players/platforms
         if (data.isLive) {